pyproject.toml
[project]
name = "zeke"
version = "0.1.0"
description = "Stateless CLI for personal knowledge management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "python-frontmatter>=1.1",
    "rich>=13.0",
]

[project.scripts]
zeke = "zeke.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes (undefined names, unused imports)
    "W",  # pycodestyle warnings
    "I",  # isort (import order)
    "B",  # flake8-bugbear (common bugs and design issues)
    "D",  # pydocstyle (enforces the google convention set below)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "frontmatter"
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
